Use firmware provided DTs
The Raspberry Pi firmware already knows how to provide and assemble device
trees that (almost) work with upstream kernels.

Add some overlays to make them work for real and change the U-Boot config
to use those instead of the built-in reference device trees.

That way, we should be much more platform agnistic than before and hopefully
able to run on any RPi device.

Signed-off-by: Alexander Graf <agraf@suse.de>